+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+* fix `LiteralValueComparisonProcessor`: fix handling for NaN values
+
 ## [1.2.0] - 2023-03-09
 
 ### Fixed